Show errors of "Change password" for local provider
When submitting the "Change your Askbot password" form previously, and
validation had an error (e.g. "Please choose password > 8 characters"),
the error was never displayed because the JavaScript always removed it.

Now, the error message is shown. The display can be improved though, as
the trailing span messes up the horizontal alignment of the input
fields.